requirements.txt
requests
pandas
feedparser
beautifulsoup4
python-dotenv